perf(cli): extract route specs from command-registry to break static import chain
The bundler hoists all transitive dependencies from static import chains
into top-level imports. route.ts statically imported findRoutedCommand from
command-registry.ts, which had 17 static imports for all command modules.
This caused 111 static imports (~860 modules, 7.56MB) to be loaded for
every CLI invocation, even simple commands like config --help.

This PR:
- Extracts route specs and findRoutedCommand into a new lightweight
  route-specs.ts that uses dynamic imports for command implementations
- Simplifies command-registry.ts by removing route-related code and 6
  heavy imports (healthCommand, statusCommand, etc.)
- Reduces the run-main bundle chunk from 111 to 59 static imports

The route specs' run() functions are already async, so the dynamic imports
add no overhead to the matching/routing logic — commands are only loaded
when a route actually matches and executes.
